requirements.txt
streamlit
openai>=1.50.0
pypdf
numpy
scikit-learn
python-dotenv
